final-project-back-2-the-future-main/code/requirements.txt
altair==5.3.0
annotated-types==0.6.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.4
docstring_parser==0.16
exceptiongroup==1.2.1
filelock==3.13.3
fsspec==2024.3.1
generativeai==0.0.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.46.0
google-cloud-bigquery==3.20.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.22.0
huggingface-hub==0.22.2
idna==3.6
importlib-metadata==6.11.0
iniconfig==2.0.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.3.0
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
PyPDF2==3.0.1
pytest==7.4.3
pytest-cov==2.11.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
shapely==2.0.3
six==1.16.0
smmap==5.0.1
streamlit==1.33.0
tenacity==8.2.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
validators==0.28.1
vertexai==1.46.0
watchdog==4.0.0
zipp==3.18.1
